[Android/ARM64] Deployment: Manipulating ELF files with patchelf
We sometimes meet a situation that have to manipulate the existing
ELF binary files on the target device because some software vendors
provide binary files only to protect their intellectual property.

In this case, the '[patchelf](https://github.com/NixOS/patchelf)'
package is helpful to us to modify the dynamic linker (ELF interpreter)
and RPATH of ELF executables and libraries as an workaround.

* What for?
 * Change the dynamic loader ("ELF interpreter") of executables
 * Change the RPATH of executables and libraries
 * Remove declared dependencies on dynamic libraries (DT_NEEDED entries)

Signed-off-by: Geunsik Lim <geunsik.lim@samsung.com>