[Android] Add Multi Device Example
This is the applicaiton for the demonstrate nnstreamer with multi
device. In order to execute this application, you may need three
devices, one for the camera input and prepocessing, one for the object
detection and the last for the pose estimation.

Once it gets the input from camera, it do the preprocessing. and after
that first device send the data to the second device. In second
device, object detection is conducted using tensorflow lite and if
there is person, then post processing is executed. after that the data
send to the last device, and the pose estimaition is done in third
device.

In order to transfer the data among devices, tcpclientsrc and
tcpserversink is used.

**Self evaluation:**
1. Build test:	 [X]Passed [ ]Failed [ ]Skipped
2. Run test:	 [X]Passed [ ]Failed [ ]Skipped

Signed-off-by: jijoong.moon <jijoong.moon@samsung.com>