[Bash/Example] Add image segmentation example with edgeTPU and TCP
Add image segmentation example with edgeTPU and TCP.
The tcp server decodes the tensor to flatbuf and transmits it.
The tcp client converts flatbuf into video and segments image in the edge tpu and display it.

Prerequisite for the server : libgstgdp.so
prerequisite for the client : libgstgdp.so, edgeTPU(device, nns_filter and runtime lib)

Signed-off-by: gichan-jang <gichan2.jang@samsung.com>