[Bash/Example/Segmentation] Improve visuals and performance
This patch makes the gst-launch image segmentation example
looks more completed by adding three things:
* Remove the background checker pattern.
* Set sync=false for videosink to improve responsiveness of the result.
* Remove the letterbox. By the limitation of the model, the model tries
to sement against the black letterbox too but it is hard to prevent it
by using available gst-launch-1.0.
So it is simpler just to cut the result image.
One potential problem is that the original video isn't guaranteed
to have a 16/9 ratio but it is a safe bet because
most webcams have 16/9 ratio by default.

Signed-off-by: Bumsik Kim <k.bumsik@gmail.com>