[native] Upload base code of face_detection_tflite and python port of object_detection_tflite
We are currently working on porting object_detection_tflite to python.
At this moment, We have implemented callbacks for tensor_sink.
One thing to note in this code, Decoding bytestring function was added to new_data_cb.
By doing so, Shared memory for boxes and detections can be read correctly as python array.
And We transformed those 1d arrays to 2d and adjusted the logics of get_detected_objects
so that it work similar to logic written in the C language.
Cairo related callbacks and debugging features will be updated.
The python file for face detection model is at base level.

Signed-off-by: JANG JONG HA <jangjongha.sw@gmail.com>